cohesive refactoring and prompt improvments
